fix a lot of 64-bit warnings

+++ source/csg/region.cpp
@@ -164,7 +164,7 @@ void Region<S, C>::AddUnique(Cube const& cube)
       // a tail-end recursive way to avoid duplicating code, but
       // I believe it would be ultimately much more confusing to 
       // read.
-      uint c = cubes_.size();
+      uint c = (uint)cubes_.size();
       while (c > 1) {
          if (!cubes_[c-2].CombineWith(cubes_[c-1])) {
             break;
@@ -207,7 +207,7 @@ void Region<S, C>::Subtract(Cube const& cube)
    CubeVector added;
 
    unsigned int i = 0;
-   unsigned int size = cubes_.size();
+   unsigned int size = (int)cubes_.size();
 
    Validate();
 
@@ -233,7 +233,7 @@ void Region<S, C>::Subtract(Cube const& cube)
    cubes_.insert(cubes_.end(), added.begin(), added.end());
    size_t addCount = added.size();
    if (addCount) {
-      _churn += added.size();
+      _churn += (int)added.size();
       CHURN_LOG(7) << "added " << added.size() << " cubes in subtract";
    }
 
@@ -328,7 +328,7 @@ template <class S, int C>
 Region<S, C> const& Region<S, C>::operator&=(Cube const& cube)
 {
    unsigned int i = 0;
-   unsigned int size = cubes_.size();
+   unsigned int size = (unsigned int)cubes_.size();
 
    Validate();
 
@@ -571,7 +571,7 @@ void Region<S, C>::OptimizeOneTagByMerge()
    if (__optimizeStrategy == WorkForward) {
       //   merged:      everything below this is fully merged.
       //   count:       number of valid cubes.  
-      uint merged = 1, c = cubes_.size();
+      uint merged = 1, c = (uint)cubes_.size();
       uint start = c;
 
       while (merged < c) {
@@ -611,7 +611,7 @@ void Region<S, C>::OptimizeOneTagByMerge()
    S areaBefore = GetArea();
 
    unsigned int i, j, k;
-   unsigned int size = cubes_.size();
+   unsigned int size = (unsigned int)cubes_.size();
    bool merged;
 
    i = 0;
@@ -864,7 +864,7 @@ Cube<S, C> Region<S, C>::GetBounds() const
    }
 
    Cube bounds = cubes_[0];
-   int i, c = cubes_.size();
+   int i, c = (int)cubes_.size();
    for (i = 1; i < c; i++) {
       bounds.Grow(cubes_[i].GetMin());
       bounds.Grow(cubes_[i].GetMax());
@@ -989,7 +989,7 @@ Region<int, C> csg::ToInt(Region<double, C> const& region) {
    Region<int, C> result;
    
    Region<double, C>::CubeVector const& cubes = region.GetContents();
-   uint i = 0, c = cubes.size();
+   uint i = 0, c = (uint)cubes.size();
 
    while (i < c) {
       Cube<int, C> icube;
